modules/kafka_consumer/requirements.txt
kafka-python
GeoAlchemy2==0.8.4
psycopg2-binary==2.8.5
